libertas_tf_usb: fix libertas_tf_usb to match structural changes in libertas_tf
The development of the libertas_tf_sdio driver introduced structural
changes in libertas_tf.  Most notably, firmware loading and starting
was needed in the probe stage due to the requirement that mac80211
have the MAC address set before the starting of mac80211 in add_card.
This patch fixes libertas_tf_usb to match these structural changes
as needed.