input: keypad: add crucial-tec optic track pad
this patch adds a new input device.
the device is connected via i2c interface to the UE, and
uses GPIO interrupt to signal movement events.
the driver uses two GPIO for reset and power off of the device.
the driver sends keypresses for up/down/left/right whenever
a movment is above a thrashhold.
the otp is used as keypad nevigation in SAAR c boards.

Change-Id: I64e10ca808b72611035fb6c6bd1ba239226fd132
Signed-off-by: Chen Reichbach creichba@marvell.com
Signed-off-by: Michael Meriin <meriinm@marvell.com>
Signed-off-by: Xiaoguang Chen <chenxg@marvell.com>