USB: OTG: add ChipIdea usb OTG driver
This driver is for ChipIdea USB OTG controller. PXA168/PXA955/PXA910/
MMP2/MMP3 all have ChipIdea USB OTG controller.

clockgating is integrated into the driver. When platform support vbus
and id pin detection or OTG is enabled, clockgating will be enabled.
So clock and phy will be on only when USB cable is plugged, and will be
off if cable is unplugged.
clock on will cost 3mA, and phy on will cost 11mA at TDDKB-A1.

Change-Id: I703c8a4c384ff342da498215762ed2a424d15aca
Signed-off-by: Chao Xie <chao.xie@marvell.com>
Signed-off-by: Neil Zhang <zhangwm@marvell.com>