reg_rw: add register access lib for some write-only bits on PXA.
Cached the bits value for the write-only bits. When reading these
bits, use the cached value and don't read from the registers.
One limitation of this lib that is every time we add one register,
one page(4KB) is used due to ioremap() even for continuous address.
So in future, if the lib is used by a lot of users, we should avoid
this mapping method to save virtual memory pages.

Change-Id: I06156498b90944ada793beedef6f2dbaabc50d3c
Signed-off-by: RaymondWu <xywu@marvell.com>