touchscreen: 88pm860x: add the touchscreen driver
It is a very ugly patch. We miss everything in commit from .32 to .35
So i have to merge the pacthes in .35.
There is no description of why we add so many code based on open source
driver. The driver own need makeit clear, and push it back to open source.
patch merged
63353fb	pxa9xx: Touch & Backlight: Enable touch and backlight
358f9aa	input: fix error detection in 88pm860x touch driver
6cac5fa	ARM: pxa: touch: fix memory leak error

Change-Id: Ib2dd1c36c2a19694893ee4965050846eaaa2400c
Signed-off-by: Chao Xie <chao.xie@marvell.com>