mmc: add dcache flush to sync I/D cache on read
In the case of data read in the mmc driver, the data may only in
the d-cache but not the real memory after the read request, like
using the bounce buffer or PIO mode. It shall cause problem if
the I-cache is not filled from d-cache or dcache aliases. Here add
rq_flush_dcache_pages after each read request to solve this issue.

Change-Id: If745f783f716ca96a84f0d62b13b15992c92b7f9
Signed-off-by: Bin Wang <binw@marvell.com>