mtd: pxa3xx_nand: minimize tr lantency
Current reading operation in newer controller adopt the naked command,
which would seperate the original one operation into several operations.

It would first send naked read command dispatch to control the NAND chip
to transfer the sepecified data to its internal ram. Then it would send
naked read sniff command to control the NAND chip to send out data from
its internal ram to controller with RE# sampling.

So in this mode, it don't need the tR parameter, for we cannot precisely
determine the very second of R/B# busy state is over. If the tR is set
small, we still wait for the READY bit to be set, while if it is set
large, we may lose much performance. So we set the tR to 0 in naked
mode, and send the following sniff command when READY bit shows up.

Change-Id: Ibe62c351bef36210ba9a3c3fbd38f0e6299134c0
Signed-off-by: Lei Wen <leiwen@marvell.com>