USB: mv_udc: add vbus event notification for charger
Changer driver shall use the exported API
	int mv_udc_register_client(struct notifier_block *nb)
to register its callback function to usb driver.

Then if the usb driver detecte a vbus connection changed event, it will
notify charger driver the charger type.
Further more, if it is an VBUS_CHARGER, it will give the current in mA
too.

Notes:
charger_work is for quick detection when vbus is down or an AC standard
charger is plug in or usb enumeration is finished.
delayed_charger_work will delay enough time for enumeration to be
finished, and this can be used to distinguish none-standard AC charger
from usb_charger.

Change-Id: Idc84173816bd342461a6a2f78109cfd49042307c
Signed-off-by: Neil Zhang <zhangwm@marvell.com>