pmem: share memory heap among pmem nodes
Allow to create a pmem node without allocated pages,
but connect it with an existing master node.

The mode can reduce the external fragmentation,
and keeps nodes with different attributes.

To use this feature, please clean the field of size
in struct android_pmem_platform_data, and set its field of 'start' to
the minor ID of pmem node what you wanna connect to

Change-Id: I0d670bcd97826e28518420d73845434986b03b38
Signed-off-by: Horace Hsieh <xiehong@marvell.com>
Signed-off-by: guoyh <guoyh@marvell.com>
Signed-off-by: Yu Tang <ytang5@marvell.com>