pmem: do not flush whole master region
If we want to use part of master region, it needs the application
to submap. However, in Marvell vmeta usage case, we cannot do that
since vemta-lib doesn't know when and how much size of the submap
area. Fix the issue by allowing to flush part of the master region
, though this is not the original design target.

Change-Id: Ib43fa8f107416c2d83ee619b8cc397c92797ef20
Signed-off-by: RaymondWu <xywu@marvell.com>
Signed-off-by: guoyh <guoyh@marvell.com>