dvb/cmmb:sms:support alloc multiple buffers
Originally, the driver use dma_alloc_coherent to alloc a large buffer
and separate it to multi small buffers for data transfering. Also the
buffer is mapped to user space by a single vma.

But it may fail very occasionally when alloc too large buffer. In order
to improve the driver's stabily, here add code to support alloc
multiple but smaller buffers.

Change-Id: I31759de801a89b2a44f98da642a5ff780cd9608b
Signed-off-by: Jialing Fu <jlfu@marvell.com>