led: add power control for 88pm860x led
kernel upgrade from .35 to 3.0;
This is led part of below power optimization patch;
589992b	88PM8606: add power optimization
Original log:
Add a voting mechanism to turn on/off the reference group and internal
oscillator. Each funtion that requires turning it on need to call
levante_ref_gp_and_osc_get() before it turns itself on and
levante_ref_gp_and_osc_release() after it turns itself off with the
enum levante_ref_gp_and_osc_clients corresponding to the field that
turns that function on.

The refernce group and internal oscillator are turned on only when
at least one of it's clients holds it on using the above mechanism.

This saves ~100uA.

Change-Id: Ic9a9bac6f1035381db774fd46d8c65e61e74459f
Signed-off-by: Jett.Zhou <jtzhou@marvell.com>