ARM: mmp: keep sensor power on during its life cycle for abilene
For abilene cannot cut off the V_SD2 which would supply the IO power of
ov5642, it disobey the ov's on/off logic, which require that when on
should first supply the core power then io power, while off require
first cut the io power then core power.

So current solution is keep regulator always on, but rely on the gpio
pin to toggle power down pin to let sensor exit/enter low power mode
when on/off the sensor.

Change-Id: I1994ff3294ece789748d7cb55202a527e032dad0
Signed-off-by: Lei Wen <leiwen@marvell.com>