ARM: conditional ioremap() fix for copy_oldmem_page()
Backporting patch from:
http://lists.infradead.org/pipermail/kexec/2011-June/005169.html

The crash kernel is loaded through kexec-tools which also contains
code that reserves memory for the elfcorehdr= option. This memory
block is reserved _inside_ the main system memory of the secondary
kernel. The /proc/vmcore code in the secondary kernel is however
using copy_oldmem_page() to access both this elfcorehdr area and
the rest of the memory used by the the first kernel.

So the copy_oldmem_page() function is used to access data that
may be located in system memory, or it may be outside. Always
using ioremap will not work, so this patch makes it conditional
based on pfn_valid().

For more details please look at the sh7372-based example here:
http://permalink.gmane.org/gmane.linux.ports.sh.devel/11502

Change-Id: I4084bef5b0746e8047d7bc42b65cffefa8943241
Signed-off-by: Magnus Damm <damm@opensource.se>
Signed-off-by: Lei Wen <leiwen@marvell.com>