kdump: add update method
This patch provides an enhancement to kexec/kdump by adding a updating
vminfo method which could be called from user space. So that while core
hang issue happen, the bootloader could directly transfer the control to
the start offset of original saved kernel relocation code entry, which
lead to bring up the crash kernel and could do the kernel dump without
the panic. (This is a must where the case that cannot call the panic,
die function since the core is hang deadly.)

Change-Id: I248d5f25b2fd250210f2d609c641c0d3c7cff448
Signed-off-by: Lei Wen <leiwen@marvell.com>