ARM: mmp: clean for dma driver
Audio DMA and memory DMA have the similiar registers'
definition; Called them as Two-channel DMA engine (TDMA),
ADMA and MDMA are two different type of TDMA; So get the
unified registers definition for them.

The target is to use dmaengine to manage these dma channels,
so far just only keep this driver and let it can co-exist
with DMAC driver.

Change-Id: I1c1f8f7b2b059d262f51d8d10c02301d6d885853
Signed-off-by: Leo Yan <leoy@marvell.com>