88pm860x: added registers cache
kernel upgrade form .35 to 3.0

Amend the commit ea9916a17e4cced09f16296a1a7d55f6e5c8a013, fix the
bootup hang of TD platform, the root cause is that i2c hardware lock is
used wrongly and deadlocked in page_in/page_out function. In kernel 3.0,
hardware lock is already included in adapter_lock so it is not needed
called here.

The I2C bus is shareable resource between APPS and COMM.
Heavy traffic on I2C bus causes collisions between APPS and COMM
(Arbitration-loss, bus-error obtained)
The I2C accesses must be minimized

There is less effective but simple general solution:
Caching PMIC-Registers in RAM,
This eliminates read/write from/to registers which are not STATIS/MEASUREMENT
and so modified by SW only.

Not all registers are cacheable, they are defined in table
pmic_noncaheable_regs[]
The table is correct for the 88pm8606 and 88pm8607 only,
for other PMICs the cache will never hit.

READ:  If register cache-hit, the data obtained from RAM but not from PMIC.
WRITE: If register cache-hit and data is same as to be written,
       only RAM is updated.

The cache is working the same also for the special 88pm8607 test-page mode

Change-Id: Ia2add90231f703c25ab4255474cd3ca8dc455ae7
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Signed-off-by: Michael Meriin <meriinm@marvell.com>
Signed-off-by: Dina Gerstein <dgerstei@marvell.com>
Signed-off-by: Xiaofan Tian <tianxf@marvell.com>
Signed-off-by: Wenzeng Chen <wzch@marvell.com>
Signed-off-by: Jett.Zhou <jtzhou@marvell.com>