USB: mv_otg: adjust clock gating solution
Consider that we need to echo something to let controller into A device
mode. We can take the vbus as the only consideration when we decide
whether it is can do clock gating.
It can save about 10mA when device is in B_idle mode for that only have
vbus detection such as TD.

Change-Id: I7e7874eaf7f15b7dc92b84e88330b32a6a917cbd
Signed-off-by: Neil Zhang <zhangwm@marvell.com>