mipsram: add support
*** This patch adds scheduler and IRQ handler MIPSRAM support on kernel 3.0

MIPSRAM is a low latency service that logs each of the following events:
 1) context switch
 2) IRQ entry
 3) PM events
The log is stored in the system in a 128KB cyclic buffer in kernel space.
In order to store this data to a file when an ASSERT occurs, we need to patch the following files in the kernel:
 1) sched.c  - To log context switches between threads, processes and workqueues (implemented in Linux as kernel threads).
 2) handle.c - To log IRQ entry

The MIPSRAM kernel driver also has a service that allows the user-space EEH code to re-map the MIPSRAM buffer to it's address space.
After an assert occurs the following occurs:
 1) Kernel EEH stops MIPSRAM buffer
 2) User-space EEH re-maps the buffer to it's user-space
 3) User-space EEH dumps the buffer to the target flash as file called "mipsram.bin"
 4) User-space EEH closes the MIPSRAM device, effectively unmapping all pages

MIPSRAM logging can be disabled or enabled using "PXA MIPSRAM monitoring" switch in Linux kernel config (kconfig)

Changes by Xiaoguang:
change ioremap_nocache to memblock_reserve, memblock_remove and then ioremap_nocache

Signed-off-by: Shuki Zanyovka <shuki.zanyovka@marvell.com>
Signed-off-by: Idan Hemdat <ihemdat@marvell.com>
Signed-off-by: Xiaoguang Chen <chenxg@marvell.com>
Signed-off-by: Lu Mao <ylmao@marvell.com>

Change-Id: I3a728f16da918954003720eb26898e32ed662b2d