ARM:pxa:power: add pending IRQ/FIQ check before entering LPM
This patch add pending IRQ/FIQ check before entering LPM. If there
is, the LPM sequence will be exit to avoid delayed interrupt
handler.

This patch is to fix a wifi/bt bug. We found that the suspend
sequence may become longer than several seconds.
This issue is because the card interrupt is happen exactly in the
LPM entry sequence in which irq is disabled.
After test, we found that the period we need to check is between
IRQ/FIQ disable and edge detect setting.

Change-Id: I09da84dd238c63ec814e2acb019fc05589944f00
Signed-off-by: Lu Mao <ylmao@marvell.com>