hwmon: apds990x: Add modified driver code
1) ALS value range is alaigned to cm3601
2) Test with android apk, value shown normally.
3) Suspend and resume function was tested.
Detail:
1) The concept that the PS sensor cannot work without the ALS sensor is
imported to the code design and change corresponding code. That means if users
try to disable the ALS sensor while the PS sensor is still enable, an error info
will pop up and no change to the ALS sensor status.
2) Proximity values are reported in CM unit.
3) Sysfs interfaces are aligned to old ALS and PS sensor.
4) als_data is used to store the luxvalue and luxvalue is in "count" unit not really
lux value now.
5) LUX_UNIT flag is used to determine whether the report value is in Lx unit or just in
"count" unit and also whether the ATIME can be changed with als polling delay set.
6) Sovle problem that enable PS sensor ==>als sensor is enabled==>disable als
sensor ==> als sensor is not disabled.
7) Change set_interval use the ms not ums
8) Suspend/resume has been tested, the register bit PON can let the device be
in or out the low power sleep mode.

Change-Id: I4b4b893ef0a97d3b27233c144e7e3821e9090e4a
Signed-off-by: Fan Wu <fwu@marvell.com>