usb:remove the rndis auto config in kernel 3.0
Rebased from Yonghong's patch commit ID: 1184f03704
Also make related change to pxa955_defconfig

In kernel 3.0, the usb rndis is not enabled by
default,so remove the interface in uboot. Or in
the boot process the auto config will wait for
a long time to check the interface available until
timeout.

Test:
build a uboot, and boot board to check if there
has a long time wait.

Change-Id: I7b6335c7fcaaaad6578e4d7c4bb689ff8bd0a5f5
Signed-off-by: Leo Song <liangs@marvell.com>