ARM:pxa:power: enable cpufreq and co-exist with legacy profiler
This patch is the cpufreq patch 1.
It is to enable cpufreq and co-exist with legacy profiler.
Now cpufreq is only a driver. To change core frequency, it only
uses disable_op/enable_op.
Next step is to control core frequency directly and remove legacy
profiler.
This patch has passed Lindent and checkpatch.pl.

Change-Id: Ie742e44d33cae279e32945022e8bf40f9927ca82
Signed-off-by: Lu Mao <ylmao@marvell.com>