hwmon: delete Read ARA in cm3213
Read Alert Response Address(ARA) and it always seems failed.
It will result in the i2c-error log when booting.
This INT seems to be useless for us.

In spec:
CM3213 INT and the command register need to be specified when power-on state occures,
and at that moment, the host may receive a non-meaningful INT interrupt signal.
For the wake up condition, in order to receive the reference level, CM3213 does not activate
the INT function after the initial 3-light data conversion. If the fourth data value
is right the interrupt is asserted by the INT pin.

Change-Id: I279567110b44b2546735663404e5869d7d518ab5
Signed-off-by: Sarah Zhang <xiazh@marvell.com>