usb: gadget: mv_udc: fix boot up hang
Case: usb driver cause system hang when bootup if only enable UDC in
kernel configuration.
The root cause is the clock will be shut down when probe finished
because of clock gating. And at the time when a gadget driver is
registered, it will call usb_gadget_probe_driver and this in turn will
call mv_udc_vbus_session. If there is no cable attached at this time,
the vbus is low, so it will call stop_activity path without clock
enabled.
This patch do the following two things:
1. Add missing active flag setting in probe function.
2. If udc is not active, jump out the stop_activity path.

Change-Id: I28215390a640f5c425c8db30e2dec354ded07ac8
Signed-off-by: Neil Zhang <zhangwm@marvell.com>