PXA95X:CGM:add wakeup from accu for core interrupt
When entering CGM we need to activate a wakeup
mechanism which will interrupt to core at the
subsystem wakeup. Currently when trying to wakeup
from external sources such as uart / keys the
subsystem is getting the wakeup but there's no interrupt
to get the core out of C2. this mechanism will
assure that upon CGM exit the ACCU will generate
such an interrupt.

Change-Id: I99452bd2952522c8f89ef0c1e3554c2562a6a0ae
Signed-off-by: Amit Sharon <asharon@marvell.com>