Bluetooth: Periodic Inquiry and mgmt discovering event
Zhen Wang found below issue and verify the fix from opensource:
By using periodic inquiry, ioctl HCIGETDEVINFO returns hci_dev_info with HCI_INQUIRY flags which is different in .35 kernel.
This is incompatibility with userspace, as BlueZ use this flag to decide either OCF_INQUIRY_CANCEL or OCF_EXIT_PERIODIC_INQUIRY to stop discovery by HCI.
As android starts the process with periodic mode, OCF_EXIT_PERIODIC_INQUIRY is expected to stop the discovery, or it will keeps inquiring.
http://www.spinics.net/lists/linux-bluetooth/msg16205.html

Open source patch commit:
By using periodic inquiry command we're not able to detect correctly
when the controller has started inquiry.

Today we have this workaround in inquiry result event handler to set
the HCI_INQUIRY flag when it sees the first inquiry result event.
This workaround isn't enough because the device may be performing
an inquiry but the HCI_INQUIRY flag is not set. For instance, if
there is no device in range, no inquiry result event is generated,
consequently, the HCI_INQUIRY flags isn't set when it should so.

We rely on HCI_INQUIRY flag to implement the discovery procedure
properly. So, as we aren't able to clear/set the HCI_INQUIRY flag in
a reliable manner, periodic inquiry events shouldn't change the
HCI_INQUIRY flag. In future, if needed, we might add a new flag (e.g.
HCI_PINQUIRY) to know if the controller is performing periodic
inquiry.

Thus, due to that issue and in order to keep compatibility with
userspace, periodic inquiry events shouldn't send mgmt discovering
events.

Change-Id: I6faec7f746815f332f3ceb32c2f4a9d3cc449bb7
Signed-off-by: Jun Nie <njun@marvell.com>