media: ov5642: update spec table
For Pre_div and Div1to2p5 map has the member of .5 value, if we keep it
like original as the round of integer number, the final calculated value
may not be accurated enough. Double the original value to be integer and
then divide it by two to get accurate result.

Change-Id: I376a6433497200b4363ddd859d0e5ff93b839a38
Signed-off-by: Lei Wen <leiwen@marvell.com>