ARM: MMP: CPUFreq: Bug fix for 3 cores support
Fix some bugs to support 3/2/1 cores CPUFreq/Hotplug

Note that if CONFIG_SMP is not defined we need to check the core
is MP core or MM core.

There is a known issue for UP FC, we need to modify uboot to make
the other two cores are in busy loop, otherwise FC will hang there.
Supposed it will be fixed on B0.

Change-Id: I9aebfb649f14e482b2293e91a80030e2d50416ec
Signed-off-by: Raul Xiong <xjian@marvell.com>