ARM: MMP: Add C2 for LPM idle
Currently there are three types of HW issues for C2:
1. Accessing VFP after C2 will hang
2. Snoop request on a core that in C2 will hang
3. GIC interrupt on a core that in C2 may hang
Also we have to config APMU_DEBUG registers to make it can be woken up
by interrupts.

So currently we just add the functions for C2 but not use it unless the
HW issue is fixed.
Suppose on B0 these issues will be fixed.

Change-Id: Iaf14a2bfd7e6e6182d8a3869bf202c9d30bd52dd
Signed-off-by: Raul Xiong <xjian@marvell.com>