ARM: mmp: core_morph: map first 2MB for hypervisor code
Reserved the first 2MB for hypervisor code, so map this
region as executable and non-cache with flag MT_MEMORY_NONCACHED;
later can execute the core swapping code in this region.

NOTE: This is a workaround for current core morphing's
implemention; need remove this if using core morphing
with normal world.

Change-Id: I2e882b0346d1c2db6b26c893d14ffe6534beb2a4
Signed-off-by: Leo Yan <leoy@marvell.com>