ARM: pxa910: improve pxa918 and pxa921 cputype identification
According to fuse usage doc, bit12~15 should be programmed to 0xf for
pxa920 and pxa918 processor, for pxa921 it is 0x3. But T3 fuse_id is
not following this rule. As we have already shipped these chips for our
customer, we need to change the chip type identification procedure in
the code. The current sequence is as below:
1. first check if it is pxa918 by bit24~25, if yes then skip next two
steps
2. then check if it is pxa921 by bit12~15, if yes then skip next step
3. default it is pxa920
These sequence is the safest bet for all chipped processors right now.

Change-Id: Idc6930bc8bcbbb23389b9427bedd382b8d1ff979
Signed-off-by: Jett.Zhou <jtzhou@marvell.com>