ARM: mmp: increase mmc dev minors to 16 on pxa910
As the partition num in the gpt partition table grows, it may exceed the
pre-defined mmc dev minors num 8, and the exceeded partitons need to use
the BLOCK_EXT_MAJOR(259) as major dev num, and a bit inconvenient for
up-layer apps. Here increase the mmc dev minors num to 16, as all the
partitions under the mmc shall use the a same major num. There is still
16 max mmc devices availabe in this case, shall be enough.

Change-Id: I92a6b0c53fd4f1825717fa89edfc510b212fe431
Signed-off-by: Qiming Wu <wuqm@marvell.com>