ASoc: add delay to get stable headset status on mmp3
add delay to get stable headset status on mmp3. during the headset
plug in, WM8994 may report a temporary state of headphone before it
recognizes it's a headset. this behavior causes inconveniennce to
audio HAL. so add 200ms time delay to double check whether a state
of headphone is true.

Change-Id: I59e504ba281aeeda1745ce5580a39c7932fd0e1d
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>