MTD: onenand_base: only report 4 bit ECC error
When correctable Ecc error happens, which may be
1 bit to 4 bits, don't report all to upper level,
only report when 4 bit correctable ecc error happens.

The reason to do this is that 1 or 2 bits ecc error may
happen frequently under stress test, if one block has 3
such correctable ecc error happens, yaffs will mark this
block as bad block, and BBT table will soon be used up,
but this block is not real bad. we can still use it.

Change-Id: I29ac3875d346e787db0715ea63ac98c60c569e91
Signed-off-by: Xiaoguang Chen <chenxg@marvell.com>