usb: gadget: mv_udc: fix full speed issue when bootup
The current solution will do enumeration when boot up with cable
attached. But android hasn't set configuration at this time, it will
cause enumeration failure and next time it will be recognized as full
speed controller. So remove pull up for android gadget driver.
Meanwhile, we should enable clock when cable is plugged in since we will
access register to judget what type of charger it is.

Change-Id: I847445a00a572083bcd1d5cb0127546d8168fa08
Signed-off-by: Neil Zhang <zhangwm@marvell.com>