ARM:pxa:power: fix constraint setting when IRQ disable
We will change the protection of OP change from IRQ disable to
mutex lock. So we must guarantee all constraint setting of running
OP not in IRQ disable segment. Or it will be kernel panic.
This patch is to move the constraint setting of running OP out of
IRQ disable segment in drivers, including I2C, onenand, vmeta and
overheating.

Change-Id: I7dca6d00d9999cdf0b17d1de7e4b31db6d95e246
Signed-off-by: Lu Mao <ylmao@marvell.com>