mmc: add specific code for io access handler
For sdhci-tavor controller, we disable the mmc controller to save
power. However, this will lead the mmc controller's registers can
not be accessed always.
The io access handler is implemented to ensure the register access
is always available. The access constraint will be released in the
next request-response loop. So there should be no power penalty.

Change-Id: I999ae779add5a41c98d817c3099e1f86cc9adef7
Signed-off-by: Raymond Wu <xywu@marvell.com>