ARM: MMP: remove jitter in plug/unplug.
When you plug HDMI cable, it generates several hpds, sometimes it even asymmetric,
e.g. normally when plug, it is off; unplug, on. But actually, when you plug,
it could be off..off..on or on..off..off..off..on, it will introduce power
constraints and twsi6 power setting abnormal.So we delay the hpd detecting,
1 sec after the last hpd generates we consider a real hpd and report to HDMI stack.
 If hpd generates within 1s, we simply cancel this work and make another delayed work.

Change-Id: I4649b0ad01a90f21073f3b1e8af70d21aed13c5b
Signed-off-by: zhangyf <zhangyf@marvell.com>
Signed-off-by: Baoyin Shan <byshan@marvell.com>