ARM: pxa9xx: soc_camera: fix digital zoom image shake
As a W/R for ov5642 zoom shift issue, we use DMA stop/start to
force a DDADR reload, but see ghost frame sometime. This is
because when stop DMA, the buffer being DMAed on is corrupted,
when restart DMA, controller fetch from current DDADR, and will
skip the corrupted frame and continue DMA on current position,
so the corrupted frame will be DQed.
Solution is do a DMA list clean, and let DMA restart from list
head before restart DMA

Change-Id: I52abb221911ccb2246f3341db39c7513fb43a54a
Signed-off-by: Jiaquan Su <jqsu@marvell.com>