ARM: pxa9xx: SOCAM: set default format before release
soc_camera will use the last format that is succuessfully applied
to sensor as the default format for next open. This is based on
assumption that if sensor can accept this format for this time,
then it's OK for next time. But this is not always right, because
sometime the driver is released for controller error, which
implies controller can't accept it.

Set host format to default sensor format, which is manually
guaranted to be 100% workable for every open, upon driver release.

Change-Id: I24d655e6d96d52e2b030d4723894c79bbb786cd7
Signed-off-by: Jiaquan Su <jqsu@marvell.com>