HWMON:CYWEE:Fix Raw data issue and add power ctrl
1.Fix Gyro raw data issues: The Gyro&G-sensors device node raw data
matrix have not been converted, this would affect the calibrator
accuracy of data.
2. Remove early_suspend related function: early suspend is no need for
this sensor. Because when android goes into early suspend the system
will also called sensors' disable through the sysfs interface.
So the power off control can be called in that handle function.
3. Power off operation is add at the tail of the sensors' probe function
to avoid "useless" power on status saved.
4. Sensors' interval_set interface is changed to: setting interval cannot
be used when system is in suspend related status.
5. Sensors' data_show sysfs interface is changed to: data cannot be
obtained when system is in suspend related status.

Change-Id: Id54772252182d1837bc9543f61eb58cfa4ac2efe
Signed-off-by: fwu <fwu@marvell.com>
Signed-off-by: gxie <gxie@marvell.com>