ARM: pxa95x: fix abu parameter issue
ABU need to enable both tx and rx DMA even
when playback only or record only mode.

If the non active channel is not configured,
we will use the other channel parameters.
But if the non active's channel has been configured
the configuration will be overwritten.

This patch fixed this issue by only using other
channel parameters when the non active channel is
not configured at all

Change-Id: Icc4a39820fb8ad9e32cad9ba5f861c595d896beb
Signed-off-by: Xiaofan Tian <tianxf@marvell.com>