ARM:mmp:lcd: update flip_buffer function
Sometimes, LCD display done interrupt maybe delayed, address update during this
delayed IRQ handler will cause this frame really displayed after one more VSYNC.
So this frame may be shown on screen, at the same time this buffer also be freed to user.
This may cause flicker issue. So if we found one buffer flipped in with only address
changed, we immediately write this address to register. This can improve this issue
as much as possible.

Change-Id: I186d6eaa6e1eed4ee6828148ea73540aa36938d9
Signed-off-by: Lisa Du <cldu@marvell.com>