ARM: mmp: keeps the ldo15 always on for abilene
before the i2c_3 proble, the ov8820 sensor needs to be fully powered
or not powered, or the i2c interface would be in an unknown state
and shall hang the i2c bus. as the VDD_IO for the sensor is always
on for the ov8820, the AVDD(LDO15) shall be kept always on too.

Change-Id: I5ca2578c07f4acf03f6a831337e5dd55286804dc
Signed-off-by: Bin Wang <binw@marvell.com>