ARM: mmp: Use timer2 as clock source.
If clock source and events are from the same time(timer0/timer1),
the timer counter would stop and restart when set next event. As
how long it takes is unpredictable, the timer accuracy would be
affected. So we use timer2 to seperate clock source and events.
It works with the same frequency of timer0(3.25M) or timer1(32k).

Change-Id: I01c42b10f6f15d3c285ff8d0c0ba3e8589d3efb9
Signed-off-by: fswu <fswu@marvell.com>