PXA968: MSPM Profiler Low PP recommendation fix
The MSPM Profiler will recommend at LPM exit to enter
The lowest PP. Currently the initialization of this
PP is incorrect.
This patch fixes this issue and looks for 156M
as the lowest PP to recommend. This will enable
to remove the D0CS OP from the PP table structure
without change in this code.

Change-Id: I5c3ffd4a92ffad60e3ef683583209bd77dbfba10
Signed-off-by: Amit Sharon <asharon@marvell.com>