sched: Cleanup cpu_active madness
This patch fixes cpu hotplug deadlock issue, which is a backport of
upstream 5fbd036b552f633abb394a319f7c62a5c86a9cd7

Issues found when doing emmc + hotplug stress:

     | cpu0 |                            | cpu1 |
        |                                   |
        V                                   |
  A: __cpu_up to wake up cpu1               |
        |                                   |
        V                                   |
  B: __cpu_up will finally call             |
      boot_secondary                        |
        |                                   |
        |------------------------------->   |
        |       gic_raise_irq               |
        V                                   |
  C: wait cpu1 set online bit               |
        |                                   V
        |<----------------------------secondary_start_kernel set online bit
        V                                   |
  D: cpu_notify with CPU_ONLINE             |
        |                                   V
        |                             wait for active bit set
        |                                   |
        V                                   |
  E: sched_cpu_active, set active bit ----> |
                                            V
                                      local_irq_enable

The problem is between state C and E, core 0 is scheduled out and then may call
smp_call_function to core 1. Then the deadlock happens. It because core 1’s irq
is stilled disabled by waiting core 0 to set its active mask and core 0 is
blocking at smp_call_function.

And the new sequence is like this:

     | cpu0 |                            | cpu1 |
        |                                   |
        V                                   |
      __cpu_up to wake up cpu1              |
        |                                   |
        V                                   |
      __cpu_up will finally call            |
      boot_secondary                        |
        |                                   |
        |------------------------------->   |
        |       gic_raise_irq               |
        V                                   |
      wait cpu1 set online bit              |
        |                                   V
        |                             secondary_start_kernel call notify_cpu_starting
	|                                   |
	|                                   |
	|                                   V
	|                             sched_cpu_active, set active bit
        |                                   |
	|                                   |
	|                                   |
	|                                   V
	|                             set online bit
        |                                   |
	|  <--------------------------------|
        V                                   |
      cpu_notify with CPU_ONLINE            |
                                            |
                                            V
                                      local_irq_enable

cpu active maks is set before online mask so there is no corner case now.

From the upstream 5fbd036b552f633abb394a319f7c62a5c86a9cd7's comments:

Now the purpose of cpu_active is mostly with bringing down a cpu, where
we mark it !active to avoid the load-balancer from moving tasks to it
while we tear down the cpu. This is required because we only update the
sched_domain tree after we brought the cpu-down. And this is needed so
that some tasks can still run while we bring it down, we just don't want
new tasks to appear.

On cpu-up however the sched_domain tree doesn't yet include the new cpu,
so its invisible to the load-balancer, regardless of the active state.
So instead of setting the active state after we boot the new cpu (and
consequently having to wait for it before enabling interrupts) set the
cpu active before we set it online and avoid the whole mess.

Change-Id: I6e8ab8f8bbbecb466ad23af65b5336bc9455381e
Signed-off-by: Raul Xiong <xjian@marvell.com>