PXA9XX: ACIPC: moving constraints to workqueue
Currently the ACIPC driver was taking constraints
for CP high freq DDR requests in interrupt context.
Since the DVFM process was divided to critical section
in which the protection is by mutex and the lower level
is protected by interrupt disabling we need to change
the ACIPC implementation so it will not call the DVFM
process from interrupt context and thus could
theoritcally create deadlock condition.

This patch moves the constraints of High freq DDR to
special workqueue of the ACIPC driver.

Currently there are cases where we still going to
idle before handling the workqueue. This patch is
a temporary solution until the above will be fixed.
This patch will allow to work with comm power enabled

Change-Id: I4cd52a6c4f28aa5580d2c7770c63ef6df906f5f0
Signed-off-by: Amit Sharon <asharon@marvell.com>