pxa9xx: ramdump remove ioremap
RAMDUMP error handling process should be free of any risky operations.
This is how it has been originally implemented.

For some reason in kernel 3.0 version an ioremap() has been added
for register 48100f10. Result: when handling oops in interrupt,
this causes BUG in mm/vmalloc.c#2199.
The ioremap here is not needed at all, because the same register is
read again in the same function using __REG (fixed mapping).

Removing the faulty code.

Change-Id: I8ede5eb5198f972cbe425d8b97a2e1d5cb39900f
Signed-off-by: Anton Eidelman <antone@marvell.com>